% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_branches.R
\name{.branch_order}
\alias{.branch_order}
\title{The parent's declared branch order}
\usage{
.branch_order(name, meta)
}
\arguments{
\item{name}{Character. Parent target name.}

\item{meta}{The \code{meta} tibble.}
}
\value{
A character vector of branch names, possibly empty.
}
\description{
The parent's declared branch order
}
\keyword{internal}
