% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_targets_data.R
\name{.empty_meta}
\alias{.empty_meta}
\title{Build a meta tibble of NAs when no store exists}
\usage{
.empty_meta(target_names)
}
\description{
Build a meta tibble of NAs when no store exists
}
\keyword{internal}
