% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_extract.R
\name{.file_defines}
\alias{.file_defines}
\title{Names of functions defined at the top level of an R file}
\usage{
.file_defines(path)
}
\arguments{
\item{path}{Path to an R file.}
}
\value{
A character vector of function names, possibly empty.
}
\description{
Names of functions defined at the top level of an R file
}
\keyword{internal}
