% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_targets.R
\name{.find_free_port}
\alias{.find_free_port}
\title{Find a free TCP port starting from the given port}
\usage{
.find_free_port(start_port)
}
\description{
Find a free TCP port starting from the given port
}
\keyword{internal}
