% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_function_pages.R
\name{.find_function_line}
\alias{.find_function_line}
\title{Find the line where a function is defined in an R file}
\usage{
.find_function_line(func_name, r_file)
}
\description{
Find the line where a function is defined in an R file
}
\keyword{internal}
