% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_target_pages.R
\name{.local_mermaid}
\alias{.local_mermaid}
\title{Generate a mermaid graph string for local dependencies (2 hops)}
\usage{
.local_mermaid(target_name, dependency)
}
\description{
Generate a mermaid graph string for local dependencies (2 hops)
}
\keyword{internal}
