% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_target_pages.R
\name{.target_repo_link}
\alias{.target_repo_link}
\title{Build a repo link to the target's definition in _targets.R}
\usage{
.target_repo_link(target_name, cfg)
}
\description{
Build a repo link to the target's definition in _targets.R
}
\keyword{internal}
