% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moxygenise.R
\name{moxygenise}
\alias{moxygenise}
\title{Generate Rd documentation from roxygen comments (whole directory)}
\usage{
moxygenise(codepath, manpath)
}
\arguments{
\item{codepath}{Character. Path to directory containing R source files.}

\item{manpath}{Character. Path to directory where \code{.Rd} files will be saved.}
}
\value{
Invisibly \code{NULL}.
}
\description{
Generate Rd documentation from roxygen comments (whole directory)
}
