% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moxygenise.R
\name{moxygenise_file}
\alias{moxygenise_file}
\title{Generate Rd documentation from a single R script}
\usage{
moxygenise_file(file, manpath)
}
\arguments{
\item{file}{Character. Path to a single \code{.R} source file.}

\item{manpath}{Character. Path to directory where \code{.Rd} files will be saved.}
}
\value{
Invisibly \code{NULL}.
}
\description{
Generate Rd documentation from a single R script
}
