% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_notes.R
\name{read_note}
\alias{read_note}
\title{Read a note file, returning empty string if not found}
\usage{
read_note(name, type = c("targets", "functions"), cfg)
}
\arguments{
\item{name}{Character. Target or function name.}

\item{type}{One of \code{"targets"} or \code{"functions"}.}

\item{cfg}{A site config list.}
}
\value{
Character string (possibly empty).
}
\description{
Read a note file, returning empty string if not found
}
