% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup_site_dirs}
\alias{setup_site_dirs}
\title{Create all required output directories}
\usage{
setup_site_dirs(cfg)
}
\arguments{
\item{cfg}{A site config list produced by \code{\link[=build_site_config]{build_site_config()}}.}
}
\value{
\code{cfg} invisibly.
}
\description{
Create all required output directories
}
