% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_targets.R
\name{view_tardoc}
\alias{view_tardoc}
\title{Open the Tier 1 static viewer (file://, no dependencies)}
\usage{
view_tardoc(project_path = ".", site_dir = "tardoc")
}
\arguments{
\item{project_path}{Path to the targets project root.}

\item{site_dir}{Subfolder used in \code{\link[=document_targets]{document_targets()}}.}
}
\description{
Open the Tier 1 static viewer (file://, no dependencies)
}
